Fix refactor
