+++ electron/utils/get-device-specs.ts
