+++ electron/utils/get-resource-paths.ts
